+++ routes/index.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const path = require('path');
-const  clog  = require('./Middleware/clog.js');
-const api = require('./routes/index.js');
+const  clog  = require('../Middleware/clog.js');
+//const api = require('./routes/index.js');
 const notes = require('./routes/notes.js');
 const fb = express.Router();
 
@@ -15,7 +15,7 @@ app.use(clog);
 // Middleware for parsing JSON and urlencoded form data
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
-app.use('/api', api);
+//app.use('/api', api);
 
 app.use(express.static('public'));
